fixed some conversion problems when compiling without vectorization support
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2927 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/parallel/datadriven/basis/linear/noboundary/operation/OperationMultipleEvalIterativeSPHybridX86SimdOCLLinear.cpp
@@ -302,7 +302,7 @@ namespace sg {
               for (size_t d = 0; d < dims; d++) {
                 float eval = ((ptrLevel[(j * dims) + d]) * (ptrData[(i * dims) + d]));
                 float index_calc = eval - (ptrIndex[(j * dims) + d]);
-                float abs = fabs(index_calc);
+                float abs = static_cast<float>(fabs(index_calc));
                 float last = 1.0f - abs;
                 float localSupport = std::max<float>(last, 0.0f);
                 curSupport *= localSupport;
@@ -587,9 +587,9 @@ namespace sg {
               for (size_t d = 0; d < dims; d++) {
                 float eval = ((ptrLevel[(j * dims) + d]) * (ptrData[(i * dims) + d]));
                 float index_calc = eval - (ptrIndex[(j * dims) + d]);
-                float abs = fabs(index_calc);
+                float abs = static_cast<float>(fabs(index_calc));
                 float last = 1.0f - abs;
-                float localSupport = std::max<double>(last, 0.0f);
+                float localSupport = std::max<float>(last, 0.0f);
                 curSupport *= localSupport;
               }
